reorganized quiz questions as preparation to make more questions
