+++ app/src/main/java/com/example/rbcfootprint/ui/learn/LearnActivity.kt
@@ -19,7 +19,7 @@ class LearnActivity : AppCompatActivity() {
 
         val button2 = findViewById<Button>(R.id.skipButton)
         button2.setOnClickListener{
-            val intent = Intent(this, LearnActivity::class.java)
+            val intent = Intent(this, ItemAdapter::class.java)
             startActivity(intent)
         }
     }
