+++ app/src/main/res/layout/activity_quiz_question.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
 
     android:background="@color/white"
-    tools:context=".ui.quiz.QuizQuestionActivity">
+    tools:context=".ui.quiz.firstquestion.QuizQuestionActivity">
 
     <Button
         android:id="@+id/answer4"
@@ -173,7 +173,7 @@
         android:id="@+id/textView5"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Question 1/3"
+        android:text="Question 2/3"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
