+++ app/src/main/res/layout/activity_quiz_wrong_ans1.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
 
     android:background="@color/white"
-    tools:context=".ui.quiz.FirstWrongQActivity">
+    tools:context=".ui.quiz.firstquestion.FirstWrongQActivity">
 
     <Button
         android:id="@+id/backButton2"
