+++ app/src/main/res/layout/activity_quiz_wrong_ans3.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
 
     android:background="@color/white"
-    tools:context=".ui.quiz.ThirdWrongQActivity">
+    tools:context=".ui.quiz.firstquestion.ThirdWrongQActivity">
 
     <TextView
         android:id="@+id/textView3"
