+++ app/src/main/res/layout/activity_quiz_wrong_ans4.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
 
     android:background="@color/white"
-    tools:context=".ui.quiz.FourthWrongQActivity">
+    tools:context=".ui.quiz.firstquestion.FourthWrongQActivity">
 
     <ImageButton
         android:id="@+id/backButton"
